backend/requirement.txt
fastapi
uvicorn[standard]
python-dotenv
httpx
sqlalchemy>=2.0
asyncpg
alembic
passlib[bcrypt]
python-jose[cryptography]
pydantic
pip install pydantic[email]
pydantic_ai
langgraph